requirements.txt
xarray
numpy
pandas
PyYAML